refactor(config): replace deep imports with public API imports (#2443)
Replace non-public API imports in src/vibe3/config/ with public API imports:
- from vibe3.clients.git_client import GitClient -> from vibe3.clients import GitClient
- from vibe3.clients.git_client import find_repo_root -> from vibe3.clients import find_repo_root
