+++ src/vibe3/config/convention_resolver.py
@@ -17,7 +17,7 @@ from vibe3.config.env_override import get_env_override
 from vibe3.config.profile_convention import ProfileConvention
 
 if TYPE_CHECKING:
-    from vibe3.clients.git_client import GitClient
+    from vibe3.clients import GitClient
     from vibe3.config.profile_config import ProfileConfig
     from vibe3.models.adapter_manifest import AdapterManifest
 
@@ -65,7 +65,7 @@ class ConventionResolver:
             GitClient instance (injected or lazy-loaded)
         """
         if self._git_client is None:
-            from vibe3.clients.git_client import GitClient
+            from vibe3.clients import GitClient
 
             self._git_client = GitClient()
         return self._git_client
